Solução completa para subir na AWS/pyproject.toml
[project]
name = "banqi-conversacional"
version = "0.1.0"
description = "Agente conversacional banQi — empréstimo consignado via WhatsApp"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.0.0,<2.0",
    "strands-agents-tools>=0.1.0,<1.0",
    "bedrock-agentcore>=1.0.0,<2.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "boto3>=1.35.0,<2.0",
    "httpx>=0.27.0,<1.0",
    "python-dotenv>=1.0.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "moto[all]>=5.0",
    "mypy>=1.10",
    "ruff>=0.8.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "S", "B", "UP", "SIM"]
ignore = [
    "S101",   # assert — aceitável em testes
    "S105",   # hardcoded secrets — falsos positivos em configurações
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["mock_api", "cod_poc"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "unit: Testes unitários (sem AWS)",
    "integration: Testes de integração com mocks",
    "e2e: Testes end-to-end contra staging",
]
addopts = "-q"
